Moved identifierName() util fun
Other rule sets should not have a dependency to the naming rule set.
