+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/naming/util/ExcludeClass.kt
@@ -1,6 +1,6 @@
 package io.gitlab.arturbosch.detekt.rules.naming.util
 
-import io.gitlab.arturbosch.detekt.rules.naming.identifierName
+import io.gitlab.arturbosch.detekt.rules.identifierName
 import org.jetbrains.kotlin.psi.KtDeclaration
 import org.jetbrains.kotlin.psi.psiUtil.containingClassOrObject
